test: add insta snapshot tests to adapter crates

+++ crates/uselesskey-aws-lc-rs/tests/snapshots_aws_lc_rs.rs
@@ -52,7 +52,7 @@ mod snapshot_tests {
             #[derive(Serialize)]
             struct RsaModulusInfo {
                 label: &'static str,
-                bits: u32,
+                bits: usize,
                 modulus_len: usize,
             }
